group redirectPage 생성
