+++ src/pages/GroupPage.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect } from "react";
-import { useParams, useNavigate } from "react-router-dom";
+import { useParams } from "react-router-dom";
 import useAuth from "../hooks/useAuth";
 import useBaseStore from "@/stores/baseStore";
 import GroupMenuBtn from "../components/group/GroupMenuBtn";
@@ -13,7 +13,6 @@ import EventDialog from "@/components/notice/EventDialog";
 
 const GroupPage: React.FC = () => {
   const { user } = useAuth();
-  const navigate = useNavigate();
 
   const { groupId: paramsGroupId } = useParams();
   const groupList = useBaseStore((state) => state.groupList);
@@ -39,18 +38,6 @@ const GroupPage: React.FC = () => {
     paramsGroupId,
     getGroup,
   ]);
-
-  useEffect(() => {
-    if (!paramsGroupId && groupList) {
-      if (groupList.length === 0) {
-        navigate("/group/new", { replace: true });
-        return;
-      } else {
-        navigate(`/group/${groupList[0].id}`, { replace: true });
-        return;
-      }
-    }
-  }, [groupList, navigate, paramsGroupId]);
 
   if (!groupList || !memberList || !targetGroup) {
     return (
